git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/data-race-detection@42345 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/pathformula/ctoformula/ExpressionToFormulaVisitor.java
@@ -110,7 +110,7 @@ public class ExpressionToFormulaVisitor
 
   private Formula getPointerTargetSizeLiteral(
       final CPointerType pointerType, final CType implicitType) {
-    final int pointerTargetSize = conv.getSizeof(pointerType.getType());
+    final long pointerTargetSize = conv.getSizeof(pointerType.getType());
     return mgr.makeNumber(conv.getFormulaTypeFromCType(implicitType), pointerTargetSize);
   }
